requirements.txt
anthropic>=0.40.0
python-telegram-bot[job-queue]>=20.0
requests>=2.31.0
python-dotenv>=1.0.0
pyzotero>=1.5.0
rich>=13.0.0
